requirements.txt
Babel==1.3
Django==1.8.2
Jinja2==2.7.3
Markdown==2.6.2
MarkupSafe==0.23
PyJWT==1.4.0
PyYAML==3.11
Pygments==2.0.2
Sphinx==1.3.1
alabaster==0.7.6
argparse==1.3.0
behave==1.2.5
defusedxml==0.4.1
dj-database-url==0.3.0
dj-static==0.0.6
django-extensions==1.5.5
django-filter==0.10.0
django-haystack==2.4.0
django-toolbelt==0.0.1
djangorestframework==3.1.3
docutils==0.12
gunicorn==19.3.0
ipython==3.2.1
jsonschema==2.5.1
mistune==0.7
nose==1.3.7
numpydoc==0.5
oauthlib==1.0.1
parse==1.6.6
parse-type==0.3.4
pbr==0.10.8
psycopg2==2.6.1
ptyprocess==0.5
python-ptrace==0.8.1
python-social-auth==0.2.12
python3-openid==3.0.6
pytz==2015.4
pyzmq==14.7.0
requests==2.7.0
requests-oauthlib==0.5.0
selenium==2.47.0
six==1.9.0
snowballstemmer==1.2.0
sphinx-rtd-theme==0.1.8
static3==0.6.1
stevedore==1.3.0
terminado==0.5
tornado==4.2.1
virtualenv==12.0.7
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.1
whitenoise==2.0.2
yolk3k==0.8.7
